+++ dotcom-rendering/src/web/browser/newsletterEmbedIframe/init.ts
@@ -1,43 +1,9 @@
 import '../webpackPublicPath';
 import { startup } from '@root/src/web/browser/startup';
+import newsletterEmbedIframe from '@guardian/common-rendering/src/newsletterEmbedIframe';
 
-type NewsletterHeightEventType = { source: { location: { href: string } } };
-
-const init = (): Promise<void> => {
-	const allIframes: HTMLIFrameElement[] = [].slice.call(
-		document.querySelectorAll('.email-sub__iframe'),
-	);
-
-	window.addEventListener('message', (event) => {
-		const iframes: HTMLIFrameElement[] = allIframes.filter((i) => {
-			try {
-				return (
-					i.src ===
-					(event as NewsletterHeightEventType).source.location.href
-				);
-			} catch (e) {
-				return false;
-			}
-		});
-		if (iframes.length !== 0) {
-			try {
-				const message: { [key: string]: string } = JSON.parse(
-					event.data,
-				);
-				switch (message.type) {
-					case 'set-height':
-						iframes.forEach((iframe) => {
-							iframe.height = message.value;
-						});
-						break;
-					default:
-				}
-				// eslint-disable-next-line no-empty
-			} catch (e) {}
-		}
-	});
-
-	return Promise.resolve();
-};
-
-startup('newsletterEmbedIframe', null, init);
+startup(
+	'newsletterEmbedIframe',
+	null,
+	newsletterEmbedIframe('.email-sub__iframe'),
+);
